# openapi.yaml
components:
  schemas:
    ArticleCreate:
      properties:
        category:
          title: Category
          type: string
        content_path:
          title: Content Path
          type: string
        creation_date:
          format: date-time
          title: Creation Date
          type: string
        modification_date:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Modification Date
        name:
          title: Name
          type: string
        publication_date:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Publication Date
        subtitle:
          anyOf:
          - type: string
          - type: 'null'
          title: Subtitle
        tags:
          items:
            type: string
          title: Tags
          type: array
        thumbnail:
          title: Thumbnail
          type: string
      required:
      - name
      - tags
      - creation_date
      - category
      - content_path
      - thumbnail
      title: ArticleCreate
      type: object
    ArticleRead:
      properties:
        category:
          title: Category
          type: string
        content_path:
          title: Content Path
          type: string
        creation_date:
          format: date-time
          title: Creation Date
          type: string
        id:
          format: uuid
          title: Id
          type: string
        modification_date:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Modification Date
        name:
          title: Name
          type: string
        publication_date:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Publication Date
        subtitle:
          anyOf:
          - type: string
          - type: 'null'
          title: Subtitle
        tags:
          items:
            type: string
          title: Tags
          type: array
        thumbnail:
          title: Thumbnail
          type: string
      required:
      - name
      - tags
      - creation_date
      - category
      - content_path
      - thumbnail
      - id
      title: ArticleRead
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: API for managing articles with details like name, tags, dates, category,
    etc.
  title: Article Management API
  version: 1.0.0
openapi: 3.1.0
paths:
  /articles/:
    get:
      operationId: get_articles_articles__get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ArticleRead'
                title: Response Get Articles Articles  Get
                type: array
          description: Successful Response
      summary: Get Articles
    post:
      operationId: create_article_articles__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleCreate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Article
  /articles/{article_id}:
    delete:
      operationId: delete_article_articles__article_id__delete
      parameters:
      - in: path
        name: article_id
        required: true
        schema:
          format: uuid
          title: Article Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Article
    get:
      operationId: get_article_articles__article_id__get
      parameters:
      - in: path
        name: article_id
        required: true
        schema:
          format: uuid
          title: Article Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Article
    put:
      operationId: update_article_articles__article_id__put
      parameters:
      - in: path
        name: article_id
        required: true
        schema:
          format: uuid
          title: Article Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleCreate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Article